Create logic.js file
